+++ tests/OneToOneTest.php
@@ -68,13 +68,9 @@ class OneToOneTest extends PHPUnit_Framework_TestCase
     public function testCreateTables()
     {
         $user = new OTO_User();    
-        $sql = "CREATE TABLE users (user_id INTEGER NOT NULL AUTO_INCREMENT,user_name VARCHAR(80) NOT NULL UNIQUE DEFAULT '',user_mail VARCHAR(80) NOT NULL,user_password VARCHAR(80) NOT NULL,PRIMARY KEY (user_id))";
-        $this->assertEquals($sql, (string) $user);
         $this->assertEquals(true, $user->createTable());
         
         $profile = new Profile();
-        $sql = "CREATE TABLE profiles (profile_id INTEGER NOT NULL AUTO_INCREMENT,user_id INTEGER NOT NULL UNIQUE,bio VARCHAR(255) NOT NULL DEFAULT '',PRIMARY KEY (profile_id),FOREIGN KEY (user_id) REFERENCES users(user_id) ON UPDATE CASCADE ON DELETE CASCADE)";
-        $this->assertEquals($sql, (string) $profile);
         $this->assertEquals(true, $profile->createTable());
     }
     
@@ -112,14 +108,19 @@ class OneToOneTest extends PHPUnit_Framework_TestCase
         $user = new OTO_User();
         $profile = new Profile();
         
+        $exceptionThrown = false;
+        
         try {
             $this->assertEquals(true, $user->dropTable());
             $this->assertEquals(true, $profile->dropTable());
         } catch (\Aurora\Error\DatabaseException $e) {
+            $exceptionThrown = true;
             $this->assertEquals(true, true);
         }
         
-        $this->assertEquals(true, $profile->dropTable());
-        $this->assertEquals(true, $user->dropTable());
+        if ($exceptionThrown) {
+            $this->assertEquals(true, $profile->dropTable());
+            $this->assertEquals(true, $user->dropTable());
+        }
     }
 }
